ms5611: build on linux (navio2)
