+++ src/drivers/barometer/ms5611/ms5611_main.cpp
@@ -1,0 +1,277 @@
+/****************************************************************************
+ *
+ *   Copyright (c) 2012-2015 PX4 Development Team. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in
+ *    the documentation and/or other materials provided with the
+ *    distribution.
+ * 3. Neither the name PX4 nor the names of its contributors may be
+ *    used to endorse or promote products derived from this software
+ *    without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
+ * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
+ * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ ****************************************************************************/
+
+/**
+ * @file ms5611.cpp
+ * Driver for the MS5611 and MS5607 barometric pressure sensor connected via I2C or SPI.
+ */
+
+#include "MS5611.hpp"
+#include "ms5611.h"
+
+enum MS5611_BUS {
+	MS5611_BUS_ALL = 0,
+	MS5611_BUS_I2C_INTERNAL,
+	MS5611_BUS_I2C_EXTERNAL,
+	MS5611_BUS_SPI_INTERNAL,
+	MS5611_BUS_SPI_EXTERNAL
+};
+
+/*
+ * Driver 'main' command.
+ */
+extern "C" __EXPORT int ms5611_main(int argc, char *argv[]);
+
+/**
+ * Local functions in support of the shell command.
+ */
+namespace ms5611
+{
+
+/*
+  list of supported bus configurations
+ */
+struct ms5611_bus_option {
+	enum MS5611_BUS busid;
+	const char *devpath;
+	MS5611_constructor interface_constructor;
+	uint8_t busnum;
+	MS5611 *dev;
+} bus_options[] = {
+#if defined(PX4_SPIDEV_EXT_BARO) && defined(PX4_SPI_BUS_EXT)
+	{ MS5611_BUS_SPI_EXTERNAL, "/dev/ms5611_spi_ext", &MS5611_spi_interface, PX4_SPI_BUS_EXT, NULL },
+#endif
+#ifdef PX4_SPIDEV_BARO
+	{ MS5611_BUS_SPI_INTERNAL, "/dev/ms5611_spi_int", &MS5611_spi_interface, PX4_SPI_BUS_BARO, NULL },
+#endif
+#ifdef PX4_I2C_BUS_ONBOARD
+	{ MS5611_BUS_I2C_INTERNAL, "/dev/ms5611_int", &MS5611_i2c_interface, PX4_I2C_BUS_ONBOARD, NULL },
+#endif
+#ifdef PX4_I2C_BUS_EXPANSION
+	{ MS5611_BUS_I2C_EXTERNAL, "/dev/ms5611_ext", &MS5611_i2c_interface, PX4_I2C_BUS_EXPANSION, NULL },
+#endif
+#ifdef PX4_I2C_BUS_EXPANSION1
+	{ MS5611_BUS_I2C_EXTERNAL, "/dev/ms5611_ext1", &MS5611_i2c_interface, PX4_I2C_BUS_EXPANSION1, NULL },
+#endif
+#ifdef PX4_I2C_BUS_EXPANSION2
+	{ MS5611_BUS_I2C_EXTERNAL, "/dev/ms5611_ext2", &MS5611_i2c_interface, PX4_I2C_BUS_EXPANSION2, NULL },
+#endif
+};
+#define NUM_BUS_OPTIONS (sizeof(bus_options)/sizeof(bus_options[0]))
+
+/**
+ * Start the driver.
+ */
+static bool
+start_bus(struct ms5611_bus_option &bus, enum MS56XX_DEVICE_TYPES device_type)
+{
+	if (bus.dev != nullptr) {
+		PX4_ERR("bus option already started");
+		return false;
+	}
+
+	prom_u prom_buf;
+	device::Device *interface = bus.interface_constructor(prom_buf, bus.busnum);
+
+	if (interface->init() != OK) {
+		delete interface;
+		PX4_WARN("no device on bus %u", (unsigned)bus.busid);
+		return false;
+	}
+
+	bus.dev = new MS5611(interface, prom_buf, bus.devpath, device_type);
+
+	if (bus.dev != nullptr && OK != bus.dev->init()) {
+		delete bus.dev;
+		bus.dev = NULL;
+		return false;
+	}
+
+	int fd = px4_open(bus.devpath, O_RDONLY);
+
+	/* set the poll rate to default, starts automatic data collection */
+	if (fd == -1) {
+		PX4_ERR("can't open baro device");
+		return false;
+	}
+
+	if (ioctl(fd, SENSORIOCSPOLLRATE, SENSOR_POLLRATE_DEFAULT) < 0) {
+		px4_close(fd);
+		PX4_ERR("failed setting default poll rate");
+		return false;
+	}
+
+	px4_close(fd);
+	return true;
+}
+
+/**
+ * Start the driver.
+ *
+ * This function call only returns once the driver
+ * is either successfully up and running or failed to start.
+ */
+static int
+start(enum MS5611_BUS busid, enum MS56XX_DEVICE_TYPES device_type)
+{
+	uint8_t i;
+	bool started = false;
+
+	for (i = 0; i < NUM_BUS_OPTIONS; i++) {
+		if (busid == MS5611_BUS_ALL && bus_options[i].dev != NULL) {
+			// this device is already started
+			continue;
+		}
+
+		if (busid != MS5611_BUS_ALL && bus_options[i].busid != busid) {
+			// not the one that is asked for
+			continue;
+		}
+
+		started = started | start_bus(bus_options[i], device_type);
+	}
+
+	if (!started) {
+		return PX4_ERROR;
+	}
+
+	// one or more drivers started OK
+	return PX4_OK;
+}
+
+static int
+info()
+{
+	for (uint8_t i = 0; i < NUM_BUS_OPTIONS; i++) {
+		struct ms5611_bus_option &bus = bus_options[i];
+
+		if (bus.dev != nullptr) {
+			PX4_INFO("%s", bus.devpath);
+			bus.dev->print_info();
+		}
+	}
+
+	return 0;
+}
+
+static int
+usage()
+{
+	PX4_INFO("missing command: try 'start', 'info',");
+	PX4_INFO("options:");
+	PX4_INFO("    -X    (external I2C bus)");
+	PX4_INFO("    -I    (intternal I2C bus)");
+	PX4_INFO("    -S    (external SPI bus)");
+	PX4_INFO("    -s    (internal SPI bus)");
+	PX4_INFO("    -T    5611|5607 (default 5611)");
+	PX4_INFO("    -T    0 (autodetect version)");
+
+	return 0;
+}
+
+} // namespace
+
+int
+ms5611_main(int argc, char *argv[])
+{
+	enum MS5611_BUS busid = MS5611_BUS_ALL;
+	enum MS56XX_DEVICE_TYPES device_type = MS5611_DEVICE;
+	int ch;
+	int myoptind = 1;
+	const char *myoptarg = NULL;
+
+	/* jump over start/off/etc and look at options first */
+	while ((ch = px4_getopt(argc, argv, "T:XISs", &myoptind, &myoptarg)) != EOF) {
+		switch (ch) {
+		case 'X':
+			busid = MS5611_BUS_I2C_EXTERNAL;
+			break;
+
+		case 'I':
+			busid = MS5611_BUS_I2C_INTERNAL;
+			break;
+
+		case 'S':
+			busid = MS5611_BUS_SPI_EXTERNAL;
+			break;
+
+		case 's':
+			busid = MS5611_BUS_SPI_INTERNAL;
+			break;
+
+		case 'T': {
+				int val = atoi(myoptarg);
+
+				if (val == 5611) {
+					device_type = MS5611_DEVICE;
+					break;
+
+				} else if (val == 5607) {
+					device_type = MS5607_DEVICE;
+					break;
+
+				} else if (val == 0) {
+					device_type = MS56XX_DEVICE;
+					break;
+				}
+			}
+
+		/* FALLTHROUGH */
+
+		default:
+			return ms5611::usage();
+		}
+	}
+
+	if (myoptind >= argc) {
+		return ms5611::usage();
+	}
+
+	const char *verb = argv[myoptind];
+
+	/*
+	 * Start/load the driver.
+	 */
+	if (!strcmp(verb, "start")) {
+		return ms5611::start(busid, device_type);
+	}
+
+	/*
+	 * Print driver information.
+	 */
+	if (!strcmp(verb, "info")) {
+		return ms5611::info();
+	}
+
+	return ms5611::usage();
+}
